[Adjustment][Maxell] migrated map components
