+++ app/globals.scss
@@ -134,3 +134,39 @@ $accent: #f7eb25;
         }
     }
 }
+
+.leaflet-container {
+    width: 100%;
+    height: 100vh;
+    position: relative;
+}
+
+.glowing-dot {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.dot {
+    width: 15px;
+    height: 15px;
+    background-color: #ff0000; /* Red color */
+    border-radius: 50%;
+    box-shadow: 0 0 10px 4px rgba(255, 0, 0, 0.8);
+    animation: pulse 1.5s infinite ease-in-out;
+}
+
+@keyframes pulse {
+    0% {
+        box-shadow: 0 0 10px 4px rgba(255, 0, 0, 0.8);
+        transform: scale(1);
+    }
+    50% {
+        box-shadow: 0 0 20px 10px rgba(255, 0, 0, 0.5);
+        transform: scale(1.2);
+    }
+    100% {
+        box-shadow: 0 0 10px 4px rgba(255, 0, 0, 0.8);
+        transform: scale(1);
+    }
+}
