+++ app/layout.tsx
@@ -1,10 +1,11 @@
 import type { Metadata } from 'next'
 import localFont from 'next/font/local'
-import './globals.scss'
 import RouteChangeHandler from '@/components/RouteChangeHandler'
-import { Suspense } from 'react'
 import FullScreenLoader from '@/components/FullScreenLoader'
+import { Suspense } from 'react'
 import { Analytics } from '@vercel/analytics/react'
+import './globals.scss'
+import 'leaflet/dist/leaflet.css'
 
 const geistSans = localFont({
     src: './fonts/GeistVF.woff',
