--- components/Map.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const Map = () => {
-    return <div>Map</div>
-}
-
-export default Map
