+++ package.json
@@ -30,19 +30,22 @@
         "dotenv": "^16.4.5",
         "driver.js": "^1.3.1",
         "framer-motion": "^11.11.9",
+        "leaflet": "^1.9.4",
         "lodash.debounce": "^4.0.8",
         "multer": "^1.4.5-lts.1",
         "multer-s3": "^3.0.1",
         "next": "15.1.6",
         "nookies": "^2.5.2",
-        "react": "19.0.0",
-        "react-dom": "19.0.0",
+        "react": "^19.0.0-rc.1",
+        "react-dom": "^19.0.0-rc.1",
+        "react-leaflet": "^5.0.0-rc.2",
         "sass": "^1.80.6",
         "zustand": "^5.0.0"
     },
     "devDependencies": {
         "@types/geojson": "^7946.0.14",
         "@types/google.maps": "^3.58.1",
+        "@types/leaflet": "^1.9.16",
         "@types/lodash.debounce": "^4.0.9",
         "@types/node": "^20",
         "@types/react": "19.0.8",
